Pretend you're a Teensy 4.1 all the time
In this configuration, a Teensy 4.1 persists its data even when it's
reprogrammed. This wasn't the case in the previous commit.

I think this is the minimal amount of changes needed for EEPROM testing
on a Teensy 4.1. I wouldn't recommend using this commit more widely,
since I may have missed a configuration.