Target AQ's imxrt-hal git repository
This replaces the `imxrt-hal` path dependency with its git repository
and branch. With this commit in place, we don't need a local copy of
imxrt-hal that's checked out at a specific branch. Instead, we express
the dependency in the package manifest.

Although this commit simplifies how we coordinate our development, this
commit may make it harder to co-develop imxrt-hal with examples in this
repo. Before this change, any local change in imxrt-hal would be
immediately reflected when building a local example. But now that
there's no path dependency, this change breaks that development
workflow.

There's a few ways to resolve this:

- Temporarily replace the git / branch settings with your local path.
  Don't check that into commits.
- Use a Cargo patch to override the dependency [1]. Don't check that in
  either.
- Push changes to the imxrt-hal branch and repo. Then, run `cargo
  update` in your local repo to realize the change.

[1]: https://doc.rust-lang.org/cargo/reference/overriding-dependencies.html

I tested this commit by building the EEPROM example. The example built
without erorrs. I'll use this commit as the foundation for our EEPROM
integration.

    cargo build --all-features --example=eeprom --target=thumbv7em-none-eabihf

I notice that other examples may not build due to the way that the
aq/local branch was developed. Addressing that is out of scope. However,
I did verify that the `can.rs` example still builds.