Add upstream EEPROM module, prototype bindings
This commit integrates the EEPROM bindings prototyped in adcbc54 (within
this fork) atop the Aqueduct Fluidics teensy4-rs main branch. This took
some diligence, since histories have diverged and experiments were
manually applied to the main branch. The gist of the integration
followed

- `git show adcbc54 --stat` to reveal the necessary EEPROM files.
- For each file ${FILE} in that list, `git checkout adcbc54 --patch --
  ${FILE}` to reveal waht needs to change in the main branch.
- Inspect and apply patches.

The end result is that, for EEPROM purposes, the main branch represents
the state at which the EEPROM module was originally developed.

Just like the rest of the BSP, the EEPROM module believes that it's
running on a Teensy 4.0. It has a 1080 byte capacity and some wear
leveling. For more information on the capacity and wear leveling
routine, consult the offical Teensy 4 software project.

This commit also removes the t40/t41 specialization code. These were
leftovers from an early EEPROM experiment, and we won't need this for
our EEPROM module.

I tested this commit by

1. `cargo clean`
2. `cargo run --all-features --target=thumbv7em-none-eabihf --example=eeprom`
3. Flashing the example to a bare Teensy 4.1.

I connected to the Teensy 4.1's USB serial interface, and waited 5
seconds. The first message that appears indicates that the EEPROM
sentinel value was not found. The example then fell into the loop that
periodically incremented the counter. This behaved as expected.

I then removed power from the Teensy 4.1, reconnected power, reconnected
to the USB serial interface, and waited 5 seconds. I observe a message
that indicated the EEPROM sentinel value was now found. This expected
behavior demonstrates that the EEPROM persists values across power
cycles on the Teensy 4.1.

Without removing the Teensy 4.1's power, I reprogrammed the EEPROM
example to the board (step 3, above). I observed that the example could
not find the sentinel value. Although curious, this is an expected
behavior; HalfKay erases all of the blocks that it believes are holding
the program image, and we're using some of those blocks for persistence.
We're OK with this behavior.

For completeness, I re-ran this test against a Teensy 4.0. The power
cycle behaviors were consistent, and the Teensy 4.0 persists its state
across reprogramming cycles. There's more discussions in an
Aqueduct-specific issue.