Move EEPROM bounds check into Rust bindings
By moving the bounds check into Rust, we can more easily detect and
error when accessing beyond the EEPROM range. E2END maintains the same
meaning as it did in the C module.

I separated the bounds check for unit testing. The unit tests
demonstrate the correct bounds check, and I also updated the hardware
example to perform an occasional, out-of-bounds access. The error is
expected.