Remove IRQ enable / disable in EEPROM module
`__enable_irq()` would be written in Rust as

    unsafe { cortex_m::interrupt::enable() }

This call is `unsafe` for good reason: it could break critical sections
that runtimes like RTIC are expected to control. A resulting race
condition would be particularly gnarly to track down.

The `__disable_irq()` would be safe, but we should generally let Rust
code -- specifically, the owner of the `Eeprom` object -- manage
interrupts if they need.

I want to emphasize that, when it was originally prototyped, I did not
review this C code for suitability in teensy4-bsp. Furthermore, I did
not review this code for suitability in the Aqueduct system. My scope
was to integrate the C code, and I'm assuming that Aqueduct is
comfortable with the license associated with this C code.